+++ configs/core.js
@@ -16,7 +16,7 @@ module.exports = {
 	/* prettier-ignore */
 	rules: {
     'linebreak-style': ['error', 'unix'],
-    'quotes': ['error', 'single'/*, { avoidEscape: true, allowTemplateLiterals: false }*/],
+    'quotes': ['error', 'single', { avoidEscape: true, allowTemplateLiterals: false }],
 		'comma-dangle': ['warn', {
 			arrays: 'always-multiline',
 			objects: 'always-multiline',
@@ -78,12 +78,14 @@ module.exports = {
     'require-await': 'warn',
     'no-async-promise-executor': 'error',
     'no-await-in-loop': 'error',
-    'no-throw-literal': 'error',
+		'no-throw-literal': 'error',
+
+		'template-tag-spacing': ['error', 'never'],
 
     // eslint-plugin-destructuring:
     'destructuring/no-rename': 'error',
     'destructuring/in-params': ['error', { 'max-params': 2 }], // Allow {items.map(({ value, label}, i) => <li key={i}/* ... */</li>)}
     'destructuring/in-methods-params': 'error',
     // See also: https://mysticatea.github.io/eslint-plugin-es/rules/no-destructuring.html
-  },
+  }
 };